[eBPF] Fix the problem that '0.0.0.0' appears in IP address(#1373)
When socket `skc_family`  is `PF_INET6`, the `skc_ipv6only = 0` indicates that the IPv4 address needs to be translated into ipv6. If we use an IPV4 address in the case of `::1`,  the address becomes `0.0.0.0`.  We need to get the real IPv6 address and pass it to rust to avoid this.