[eBPF] Avoid meaningless tracking of client sockets
Socket A actively sends a request as a client (traceID is 0), and associates socket B with the thread ID.
Socket B receives a response as the client, create new traceID as the starting point for the entire tracking process.
There is a problem in tracking down like this, and a closed loop cannot be formed. This is due to receiving a response
from socket B to start the trace, but not being able to get a request from socket B to finish the entire trace.

(socket A) -- request ->
            |
          (socket B) <- response (traceID-1) ------------- The starting point of trace
                  |
                (socket C) -- request -> (traceID-1)
                       |
                     (socket D) <- response (traceID-2)
                            |
                          (socket E) -- request -> (traceID-2)
                               ... ...  (Can't finish the whole trace)

In order to avoid invalid association of the client, the behavior of creating a new trace on socket B is cancelled.

                   (socket A) ------- request -------->
                              |
                           thread-ID
                              |
                           (socket B) <---- response (Here, not create new trace.)