[eBPF] Fast match for protocol inference (#1739)
One thread or process processes the application layer data, and the protocol
inference program has successfully concluded that the protocol is A, then
this thread or process will probably process the data of protocol A later.
We can add a cache for fast matching, use the process-ID/thread-ID
to query the protocol recorded in the cache, and match the protocol preferentially.
If the match fails, a slow match is performed (all protocol sequence matches).

Due to the limitation of the number of eBPF instruction in kernel,
this feature is suitable for Linux5.2+