[eBPF] Remove duplicate data caused by MSG_PEEK
The recv() function usually takes a last argument of 0, which means fetch from the buffer, whereas MSG_PEEK means just look at the data, not fetch the data. As a result, the ebpf gets a lot of duplicate data.

We made a judgment at sys_enter_recvfrom/sys_recvmmsg/sys_recvmsg, not to collect data with MSG_PEEK.