[eBPF] Fix data loss due to 4-bytes read/write
The MySql/Kafka has two parts to read data:
The first part reads 4 bytes (generally the length), and the second part is the data.

Since the application layer read/write 4 bytes of data, it may not be mysql/kafka, which logically causes data loss. The fix here is for MySql/Kafka only and does not interfere with other applications.