[Querier] use connection pool
In the test environment, we constructed five kinds of query pressure
with QPS ranging from 16 to 56, and each query pressure lasted for 300
seconds. During the pressure test, we found two problems: 1) There will
be thousands of connections inside the deepflow-server Pod ClickHouse
Socket, and 90% of the status is TIME_WAIT; 2) When the QPS pressure
suddenly increases, it will almost certainly lead to write failure, and
the reason for the failure is that deepflow-server disconnects the
socket written to ClickHouse.

After investigation, it was found that the write failure when QPS
changed was caused by Querier using a short connection to query
ClickHouse. In addition, a large number of TIME_WAIT sockets were also
related to unreasonable MaxOpenConns/MaxIdleConns values.