[eBPF] Trigger inference service address adjustment
The address of the inference service is currently set to 0.0.0.0:54583. It is possible that a local IP address (not the loopback address 127.0.0.1) may be using port 54583, resulting in a bind failure. To avoid this possibility, the address of the inference service should be changed to 127.0.0.1:54583. Additionally, the log output should be updated accordingly.