[eBPF] Adjust the definition of process IDs in the profile metadata
Modify the metadata of the profile to use process IDs for user-space programs
and kernel thread IDs for kernel-space programs. The purpose is to accurately
track profile data. If kernel-space process IDs are used for tracking, the data
would be attributed to kernel processes such as 0 or 2, making it difficult to
discern the performance consumption of individual kernel threads. For user-space
processes, the design is to track at the process level (i.e., the flame graph
displays process names rather than thread names).