[eBPF] Use epoll() interface instead of poll()
poll() uses a polling mechanism, which means that after calling poll(), it iterates through all file descriptors to check if they are ready. Consequently, even if only a few file descriptors are ready, it will still check all of them, leading to CPU resource wastage.

On the other hand, epoll() uses an event-driven approach. When epoll_wait() is called, it only returns the ready file descriptors instead of iterating through all of them. As a result, in high-concurrency scenarios, epoll() processes only the ready file descriptors, avoiding the need for a full iteration and saving significant CPU resources.

In poll(), each time you make a system call to check for ready file descriptors, the kernel needs to copy the entire set of file descriptors and their associated data (if any) from user-space to kernel-space and back again. This data copying can become a significant overhead, especially when dealing with a large number of file descriptors or large data chunks.

On the other hand, epoll() uses a more efficient mechanism to avoid this data copying. It uses a memory-mapped data structure to directly share information between user-space and kernel-space, thus eliminating the need for redundant data copying during system calls. This makes epoll() more efficient in scenarios with a large number of file descriptors and high data throughput.