[eBPF] Java symbols table update
When an 'unknown' Frame is encountered during the symbolization process of
the Java process, it will be delayed for a fixed time (if the 'unknown' f-
rame is encountered again during this period, it will be ignored) to upda-
te the Java symbol table.

The update cycle of the Java process symbols table is 300 seconds.