[eBPF] Remove same socket no-tracing logic (#4301)
If the same socket is tracked, multiple tracking trees will appear. In order to avoid this situation,
the previous operation was to not track data on the same socket. But the disadvantage of doing this
is that some necessary tracking is truncated. Because we currently have a branch pruning function that
can avoid multiple trees, here we re-enable the same socket tracking function.