[eBPF] Add java symbols write space limit
'java_syms_space_limit' The maximum space occupied by the Java symbol files
in the target POD. Its valid range is [2, 10], which means it falls within
the interval of 2Mi to 10Mi. If the configuration value is outside this
range, the default value of 10(10Mi), will be used.

The size of the Java perf-PID.map file, excluding the additional space occupied
by 'agent.so' and log files, represents the actual limit of space occupied in the
target POD's root path ('/'). The actual writing limit is controlled by the global
variable 'g_java_syms_write_bytes_max', and the size (in bytes) of the Java perf-PID.map
file will not exceed it.