[eBPF] To correct an error in online CPU parsing (#4633)
The content of the file "/sys/devices/system/cpu/online" is generally in the format "0-CPUMAX," for example, "0-23." However, if it appears as "0-2,4-23"(use command "echo 0 > /sys/devices/system/cpu/cpu3/online" turned off CPU3), this situation would result in an error (only parsing "0-2").

This patch is intended to fix this issue.