[eBPF] Add enable flag for profiler & modify stackmap cleanup logic
Add a startup switch to the profiler to ensure that all preparations are completed before enabling the eBPF profiler to start working. This prevents premature eBPF data collection, as the user-space program may not be ready, avoiding the occupation of numerous stack map entries.

The cleanup of the stack map is consolidated and performed after each iteration to avoid instances where certain entries might be overlooked during the cleanup process.