[eBPF] Correcting MongoDB protocol inference
The MongoDB protocol involves two reads in the receiving direction for
a single request. Initially, the first read retrieves the first 16
bytes (length of the protocol header), followed by a subsequent read
to obtain the remaining data. Therefore, it is necessary to pre-store
the content from the first read and merge it with the data from the
second read before sending it to the agent.