[eBPF] Add TCP sequence number offset for HTTP/2 protocol
Due to differences in the data captured through the `af_packet` and
`eBPF methods` for HTTP/2, for example:

- Data captured using the af_packet method:
  `PING[0], HEADERS[86125]: 200 OK, DATA[86125]`

- Data captured using the eBPF method:
  `HEADERS[86125]: 200 OK, DATA[86125]`

Furthermore, both sides are unaware of the differences in the captured data.
This inconsistency can lead to inconsistent `tcpseq` values, making it chal-
lenging to correlate the data. To address this issue, it is agreed that both
methods adjust the `tcpseq` to the starting position of the first `HEADER`.