[Controller] move agent_group_config to upper directory
The agent_group_config and agent modules have a closer relationship to
business logic. Previously, they were placed in a deep directory
structure under controller/model/, making it difficult for people to
find them. Now, we're moving them to a shallower directory.

The best solution would actually be to move them outside of the server
directory.