[eBPF] Remove logic to not trace data sent from the client as the origin
The reason for removal:

1 We already have tracing tree pruning functionality, which can prune invalid traces based on
the time range of the root node (span).

2 The following scenario requires the removal of the non-tracing logic:

```
                 Requset threadID   Response threadID   Requset traceID   Response traceID
------------------------------------------------------------------------------------------
server process   TID-A              TID-B               traceID-A         0 [3]
client process   TID-B              TID-B               0 [1]             0 [2]
```
As we can see, because the client process serves as the starting point for sending data, its
request traceID is 0 (as indicated by [1]), and due to previous non-tracing logic, the value
at [2] is also 0. Consequently, the data from the server and the client cannot be associated.
By removing the non-tracing logic, [2] and [3] will generate a non-zero valid traceID, enabling
the association of the two sets of data.