fix: agent - eBPF DNS cannot obtain network tuple data (#7131)
The `sendto` functions are generally used in UDP protocols, but can also be used
in TCP after the connect function is called. `sendto()` use the datagram method
to transmit data.
In the connectionless datagram socket mode, since the local socket has not
established a connection with the remote machine, the destination address should
be specified when sending data. The sendto() function prototype is:

`int sendto(socket s, const void *msg, int len, unsigned int flags, const
            struct sockaddr *to, int tolen);`

The sendto() function has two more parameters than the send() function. The "to"
parameter specifies the IP address and port number information of the destination
machine.

Our current logic is as follows: network tuple information (IP, PORT) is obtained
by reading the corresponding fields of the kernel structure 'struct sock_common'.
Since the IP address and port are specified in the sendto() system calls,
the tuple data will not be populated into the kernel structure 'struct sock_common'.
As a result, we cannot obtain the tuple information. Therefore, when entering these
types of system calls, we need to save this information beforehand.