fix: agent - eBPF Adjust Java syms-cache update logic & error log output (#7291)
Previously, the update of the Java symbol cache depended on the new symbol file being larger in size than the existing file, which was illogical because it might not account for events like 'CompiledMethodUnload' that could reduce the number of Java functions (methods). This modification removes that check. Additionally, for the runtime of deepflow-jattach during the generation of Java symbol files, if exceptions occur, log outputs have been added.

Conflicts:
	agent/src/ebpf/user/profile/java/gen_syms_file.c