feat: agent - eBPF Improve aggregation efficiency of stack-trace string
It is possible that multiple threads of a process (or the process itself) use the same
task name (kernel structure: 'task_struct->comm'). To improve the aggregation efficiency
of function stack tracing strings, we use a unique value to fill 'kvp->k.pid' for threads
with the same name.

Conflicts:
	agent/src/ebpf/user/proc.h
	agent/src/ebpf/user/profile/profile_common.c