fix: agent - eBPF Correct the maximum data push delay (#8703)
When a delay occurs in the eBPF data push cycle (i.e., the time difference between
the current and previous trigger exceeds 100 milliseconds), we set the maximum data
push delay to a special value (`PUSH_DELAY_EXCEEDED_MARKER`: 199000) to indicate that
a delay has occurred, instead of providing the specific delay value. The reason is
that in a multi-core environment, direct variable assignment in eBPF cannot ensure
atomicity. Therefore, we use a special value to signify the occurrence of a delay.