fix: agent - eBPF Fix find_pid_by_name() missing closedir() (#8908)
This does not affect memory because it is used to check for the special case of two
`deepflow-agent` instances being started. If two agents are started, the newly launched
one will exit (releasing memory automatically).

If only one agent is started, it will properly call `closedir()` without causing any impact.