fix: agent - eBPF Fix TCP DNS client request loss (#9220)
Note that TCP DNS adds two length bytes at the beginning of the protocol,
whereas UDP DNS does not. We need to handle this properly to ensure that
these two length bytes are not sent to the upper layer.

When receiving data, the client does not first receive two bytes but instead
receives everything at once; whereas the server receives two bytes (length) first
and then receives the remaining bytes.

When the client sends a request, it combines both 'A' and 'AAAA'
type queries into a single request to the CoreDNS server. The first
two bytes represent the length, but this length only includes the
'A' query, not the combined length of both the 'A' and 'AAAA' queries
(the total size is referred to as "count" here). As a result, the
length check may miss this case.

This fixes the issue where the client fails to retrieve the request by adding direction judgment.