feat: agent - eBPF Get the remote IP and port of the DNS (#9276)
For the `sendmsg()/recvmsg()` system interfaces, the remote IP address
and port may be specified through the parameter `struct user_msghdr __user *msg`
and may not be recorded in the kernel `sock` structure (as is common in the UDP
protocol). We extract the values from the system call parameters to obtain this
data and populate the network tuple.