feat: agent - eBPF Obtain the mount point for file
To retrieve the full file path, we split the task into two parts:

1. **Use eBPF to obtain the file path and device number.**
2. **Use a PID-based cache to retrieve mount information**, specifically by caching the contents of `/proc/[pid]/mountinfo`.

Using the device number obtained via eBPF, we look up the corresponding mount point and mount source from the cache and process the data as follows:

The format of each line in `mountinfo` is:

```
[ID] [ParentID] [major:minor] [fs_root] [mount_point] [options] - [fs_type] [mount_source] [fs_options]
```

Bind mounts are filtered out because, in such cases, the file path retrieved by eBPF starts directly from `[fs_root]`, which is already a valid host path. There is no need to translate it into the container’s mount point path.

**Example:**
For the line:

```
1729 1710 253:0 /var/lib/mysql /bitnami/mysql rw,relatime - xfs /dev/mapper/centos-root rw,attr2,inode64,noquota
```

A file path like `/var/lib/mysql/xxxx.data` obtained via eBPF is already a host path and should **not** be translated to `/bitnami/mysql/xxxx.data`.

If the mount source is an NFS path, we perform path translation as follows:

* `str1` = NFS source, e.g., `"10.33.49.27:/srv/nfs/data"`
* `str2` = Path obtained from eBPF, e.g., `"/nfs/data/file1"`
* `str3` = Mount point, e.g., `"/mnt/nfs"`

We **replace the longest suffix of `str1` that matches a prefix of `str2`** with the new mount point (`str3`).

**Result:**
`"/mnt/nfs/file1"`

For other mount types (e.g., `proc`, `sysfs`, `tmpfs`, etc.), we combine the mount point with the eBPF-reported relative path:

**Final file path = mount point + path obtained from eBPF**