feat: agent - eBPF Obtain the mount point for file (#10057) (#10098)
To retrieve the full file path, we split the task into two parts:

1. **Use eBPF to obtain the file path and device number.**
2. **Use a Mountns-based cache to retrieve mount information**.

Using the device number obtained via eBPF, we look up the corresponding mount point and mount source from the cache and process the data as follows:

### 1. **Skip bind mounts**

The format of each line in `mountinfo` is:

```
[ID] [ParentID] [major:minor] [fs_root] [mount_point] [options] - [fs_type] [mount_source] [fs_options]
```

Bind mounts are filtered out because, in such cases, the file path retrieved by eBPF starts directly from `[fs_root]`, which is already a valid host path. There is no need to translate it into the container’s mount point path.

**Example:**
For the line:

```
1729 1710 253:0 /var/lib/mysql /bitnami/mysql rw,relatime - xfs /dev/mapper/centos-root rw,attr2,inode64,noquota
```

A file path like `/var/lib/mysql/xxxx.data` obtained via eBPF is already a host path and should **not** be translated to `/bitnami/mysql/xxxx.data`.

### 2. **Handle NFS mounts**

If the mount source is an NFS path, we perform path translation as follows:

* `str1` = NFS source, e.g., `"10.33.49.27:/srv/nfs/data"`
* `str2` = Path obtained from eBPF, e.g., `"/nfs/data/file1"`
* `str3` = Mount point, e.g., `"/mnt/nfs"`

**Result:**
`"10.33.49.27:/srv/nfs/data/file1"`

### 3. **Handle other mount types**

For other mount types (e.g., `proc`, `sysfs`, `tmpfs`, etc.), we combine the mount point with the eBPF-reported relative path:

**Final file path = mount point + path obtained from eBPF**

### 4. **Add periodic checks for process cache and root mount information changes.**

By default, changes in process information is checked every 5 minutes and reflected in the cache.
The mount information change check interval is 20 seconds.

Added periodic checks for process and mount information caches:

* **Periodic check of the process information cache**:
Stale process records are detected and cleaned up. These stale entries may remain if the process terminated due to OOM (Out Of Memory) or segmentation fault, and the exit was not captured by the process exit hook.
Meanwhile, the mount information associated with each process is also checked, and if any changes are detected, the mount data will be updated accordingly.

* **Periodic check of the mount information cache**:
  The system updates the cached mount info of the host's root mount namespace if it has changed.
  For container mount namespaces, their corresponding cached mount info is automatically released when the container process exits.