feat: agent - eBPF Support Rocky Linux 5.14
When loading the `socket-trace-bpf-linux-5.2_plus` eBPF bytecode on **Rocky Linux kernel 5.14**, the kernel verifier reports that the instruction limit has been exceeded. The error message is as follows:

```
BPF program is too large. Processed 1000001 insn
processed 1000001 insns (limit 1000000) max_states_per_insn 60 total_states 48637 peak_states 4495 mark_read 85

[2025-10-22T03:58:39.026Z INFO  socket_tracer::ebpf] [eBPF] WARN func load_obj__progs() [user/load.c:634] bcc_prog_load() failed. name: df_T_exit_sendmsg, Argument list too long errno: 7
```

This issue has **not been observed on other kernel versions**, and it is considered a **bug in the eBPF verifier** of the 5.14 kernel.
To work around this problem, we **load the `socket-trace-bpf-linux-common` eBPF bytecode** instead, which reduces the number of instructions and avoids exceeding the verifier limit.